chore: js client checkin using name

+++ packages/js/src/server/checkins-manager/client.ts
@@ -5,12 +5,14 @@ import { CheckinResponsePayload } from './types';
 export class CheckinsClient {
   private readonly BASE_URL = 'https://app.honeybadger.io'
   private readonly cache: Record<string, Checkin[]>
+  private readonly config: { personalAuthToken: string; logger: Types.Logger }
   private readonly logger: Types.Logger
   private readonly transport: Types.Transport
 
-  constructor(transport: Types.Transport, logger: Types.Logger) {
+  constructor(config: { personalAuthToken: string; logger: Types.Logger }, transport: Types.Transport) {
     this.transport = transport
-    this.logger = logger
+    this.config = config
+    this.logger = config.logger
     this.cache = {}
   }
 
@@ -19,8 +21,13 @@ export class CheckinsClient {
       return this.cache[projectId]
     }
 
+    if (!this.config.personalAuthToken || this.config.personalAuthToken === '') {
+      throw new Error('personalAuthToken is required')
+    }
+
     const response = await this.transport.send({
       method: 'GET',
+      headers: this.getAuthHeader(),
       endpoint: `${this.BASE_URL}/v2/projects/${projectId}/check_ins`,
       logger: this.logger,
     })
@@ -31,19 +38,20 @@ export class CheckinsClient {
     }
 
     const data: { results: CheckinResponsePayload[] } = JSON.parse(response.body)
-    const checkins = data.results.map((checkin) => {
-      const result = Checkin.fromResponsePayload(checkin)
-      result.projectId = projectId
-      return result
-    })
+    const checkins = data.results.map((checkin) => Checkin.fromResponsePayload(projectId, checkin))
     this.cache[projectId] = checkins
 
     return checkins
   }
 
   public async get(projectId: string, checkinId: string): Promise<Checkin> {
+    if (!this.config.personalAuthToken || this.config.personalAuthToken === '') {
+      throw new Error('personalAuthToken is required')
+    }
+
     const response = await this.transport.send({
       method: 'GET',
+      headers: this.getAuthHeader(),
       endpoint: `${this.BASE_URL}/v2/projects/${projectId}/check_ins/${checkinId}`,
       logger: this.logger,
     })
@@ -54,15 +62,20 @@ export class CheckinsClient {
     }
 
     const data: CheckinResponsePayload = JSON.parse(response.body)
-    const checkin = Checkin.fromResponsePayload(data)
+    const checkin = Checkin.fromResponsePayload(projectId, data)
     checkin.projectId = projectId
 
     return checkin
   }
 
   public async create(checkin: Checkin): Promise<Checkin> {
+    if (!this.config.personalAuthToken || this.config.personalAuthToken === '') {
+      throw new Error('personalAuthToken is required')
+    }
+
     const response = await this.transport.send({
       method: 'POST',
+      headers: this.getAuthHeader(),
       endpoint: `${this.BASE_URL}/v2/projects/${checkin.projectId}/check_ins`,
       logger: this.logger,
     }, { check_in: checkin.asRequestPayload() })
@@ -73,15 +86,20 @@ export class CheckinsClient {
     }
 
     const data: CheckinResponsePayload = JSON.parse(response.body)
-    const result = Checkin.fromResponsePayload(data)
+    const result = Checkin.fromResponsePayload(checkin.projectId, data)
     result.projectId = checkin.projectId
 
     return result
   }
 
   public async update(checkin: Checkin): Promise<Checkin> {
+    if (!this.config.personalAuthToken || this.config.personalAuthToken === '') {
+      throw new Error('personalAuthToken is required')
+    }
+
     const response = await this.transport.send({
       method: 'PUT',
+      headers: this.getAuthHeader(),
       endpoint: `${this.BASE_URL}/v2/projects/${checkin.projectId}/check_ins/${checkin.id}`,
       logger: this.logger,
     }, { check_in: checkin.asRequestPayload() })
@@ -95,8 +113,13 @@ export class CheckinsClient {
   }
 
   public async remove(checkin: Checkin): Promise<void> {
+    if (!this.config.personalAuthToken || this.config.personalAuthToken === '') {
+      throw new Error('personalAuthToken is required')
+    }
+
     const response = await this.transport.send({
       method: 'DELETE',
+      headers: this.getAuthHeader(),
       endpoint: `${this.BASE_URL}/v2/projects/${checkin.projectId}/check_ins/${checkin.id}`,
       logger: this.logger,
     })
@@ -104,6 +127,12 @@ export class CheckinsClient {
     if (response.statusCode !== 204) {
       this.logger.debug(`Failed to remove checkin[${checkin.name}] for project[${checkin.projectId}]: ${response.body}`)
       throw new Error(`Failed to remove checkin[${checkin.name}] for project[${checkin.projectId}]`)
+    }
+  }
+
+  private getAuthHeader() {
+    return {
+      'Authorization': `Basic ${Buffer.from(`${this.config.personalAuthToken}:`).toString('base64')}`
     }
   }
